pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "sigil"
version = "1.0.0"
description = "SIGIL - SE050 Hardware Bitcoin Wallet"
readme = "docs/README.md"
requires-python = ">=3.7"
license = {text = "MIT"}

dependencies = [
    "flask>=2.0.0",
    "gunicorn>=21.0.0",
    "requests>=2.28.0",
    "pysocks>=1.7.0",
    "qrcode>=7.0",
    "pillow>=9.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[project.scripts]
sigil-web = "sigil.web:main"
sigil-wallet = "sigil.cli.main:main"
sigil-server = "sigil.server.app:main"
sigil-client = "sigil.client.cli:main"
sigil-gui = "sigil.gui.desktop:main"

[tool.setuptools.packages.find]
include = ["sigil*"]
